model/Yangsan_model/requirements.txt
faiss_cpu==1.9.0.post1
langchain==0.3.18
langchain_core==0.3.35
langchain_openai==0.3.5
numpy==2.2.2
openai==1.62.0
pandas==2.2.3
Pillow==11.1.0
python-dotenv==1.0.1
scikit_learn==1.6.1
sentence_transformers==3.4.1
openpyxl
instagrapi